move resource-manager bootstrappers to `pkg`
